Optimized code and changed log paths for better management

+++ trunk/bin/html/showresource.php
@@ -70,53 +70,87 @@ include_once "header.inc.php";
 	elseif ($action == "Install")
 	{
 		// Start installation process by Launching the script which provides the summary of the build process 
-		echo "<script> document.location.href = 'workerapp.php?id=$id&submenu=$row[submenu]&icon=$row[icon]&action=showInstallLog#myfix'; </script>";
+		echo "<script> document.location.href = 'workerapp.php?id=$id&name=$row[name]&submenu=$row[submenu]&icon=$row[icon]&action=showInstallLog#myfix'; </script>";
 		
 		// Clear logs and scripts
-		if(is_dir("$workpath/apps/dload")) {
-			system_call("rm -rf $workpath/apps/dload/*");
+		if(is_dir("$workpath/logs/apps")) {
+			system_call("rm -rf $workpath/logs/apps/*");
 		}
 		
 		// Download app
-		appsLoader("$row[menu]","$row[submenu]");
+		appsLoader("$row[menu]","$row[name]");
 	}
 	elseif ($action == "Patch")
 	{
+		$fixLogPath = "$workpath/logs/fixes";
+		
+		// Clear logs and scripts
+		if(is_dir("$fixLogPath")) {
+			system_call("rm -rf $fixLogPath/*");
+		}
+		
+		// create log directory if not found
+		if(!is_dir("$workpath/logs")) {
+			system_call("mkdir $workpath/logs");
+		}
+		if(!is_dir("$fixLogPath")) {
+			system_call("mkdir $fixLogPath");
+		}
+		
 		echo "<div class='pageitem_bottom'\">";	
 		echo "<ul class='pageitem'>";
 
 		switch ($row[name]) {
 		
 			case "AppleIntelCPUPowerManagementPatch":
-				patchAppleIntelCPUPowerManagement();
+				patchAppleIntelCPUPowerManagement("myHack");
 			break;
 			
 			case "":
 			break;
 		}
 		
-		echo "<img src=\"icons/big/success.png\" style=\"width:80px;height:80px;position:relative;left:50%;top:50%;margin:15px 0 0 -35px;\">";
-		echo "<b><center> Patch finished.</b><br><br><b> You can now reboot the sysem to see the patch in action.</center></b>";
-		echo "<br></ul>";
+		if (is_file("$fixLogPath/success.txt")) {
+			echo "<img src=\"icons/big/success.png\" style=\"width:80px;height:80px;position:relative;left:50%;top:50%;margin:15px 0 0 -35px;\">";
+			echo "<b><center> Patch finished.</b><br><br><b> You can now reboot the sysem to see the patch in action.</center></b>";
+			echo "<br></ul>";
+			
+			echo "<b>Log:</b>\n";
+			echo "<pre>";
+			if(is_file("$workpath/logs/build/build.log"))
+				include "$workpath/logs/build/build.log";
+			echo "</pre>";
+		}
+		else {
+			echo "<img src=\"icons/big/fail.png\" style=\"width:80px;height:80px;position:relative;left:50%;top:50%;margin:15px 0 0 -35px;\">";
+			echo "<b><center> Patch failed.</b><br><br><b> Check the log for the reason.</center></b>";
+			echo "<br></ul>";
+			
+			echo "<b>Log:</b>\n";
+			echo "<pre>";
+			if(is_file("$fixLogPath/fix.log"))
+				include "$fixLogPath/fix.log";
+			echo "</pre>";
+		}
+		echo "</div>";
 	}
 
 function appsLoader($categ, $fname) {
 		global $workpath, $edp;
     	  
-    	$applogPath = "$workpath/apps/dload";
+    	$appsLogPath = "$workpath/logs/apps";
     	  	
     	if(!is_dir("$workpath/apps")) {
 			system_call("mkdir $workpath/apps");
 		}
-		if(!is_dir("$applogPath")) {
-			system_call("mkdir $applogPath");
+		
+		// create log directory if not found
+		if(!is_dir("$workpath/logs")) {
+			system_call("mkdir $workpath/logs");
 		}
-		if(!is_dir("$applogPath/statFiles")) {
-			system_call("mkdir $applogPath/statFiles");
+		if(!is_dir("$appsLogPath")) {
+			system_call("mkdir $appsLogPath");
 		}
-			
-		$createStatFile = "cd $applogPath/statFiles; touch $fname.txt";	
-		$endStatFile = "cd $applogPath/statFiles; rm -rf $fname.txt";
 		
 		//
 		// Download apps from SVN
@@ -125,16 +159,16 @@ function appsLoader($categ, $fname) {
 		$svnpath = "apps/$categ/$fname";
 			
 		if (is_dir("$packdir")) {
-			$checkoutCmd = "if svn --non-interactive --username edp --password edp --quiet --force update $packdir/$fname; then echo \"$fname file(s) updated finished<br>\" >> $applogPath/appInstall.log; touch $applogPath/success.txt; else echo \"$fname file(s) update failed (may be wrong svn path or no internet)<br>\" >> $applogPath/appInstall.log; touch $applogPath/fail.txt; fi";
+			$checkoutCmd = "if svn --non-interactive --username edp --password edp --quiet --force update $packdir/$fname; then echo \"$fname file(s) updated finished<br>\"; touch $appsLogPath/Success_$fname.txt; else echo \"$fname file(s) update failed (may be wrong svn path or no internet)<br>\"; touch $appsLogPath/Fail_$fname.txt; fi";
 
-			writeToLog("$applogPath/$fname.sh", "$createStatFile; $checkoutCmd; $endStatFile;");
-			system_call("sh $applogPath/$fname.sh >> $applogPath/appInstall.log &");
+			writeToLog("$appsLogPath/$fname.sh", "$checkoutCmd;");
+			system_call("sh $appsLogPath/$fname.sh >> $appsLogPath/appInstall.log &");
 		}
 		else {
-			$checkoutCmd = "mkdir $packdir; cd $packdir; if svn --non-interactive --username osxlatitude-edp-read-only --quiet --force co http://osxlatitude-edp.googlecode.com/svn/$svnpath; then echo \"$fname file(s) download finished<br>\" >> $applogPath/appInstall.log; touch $applogPath/success.txt; else echo \"$fname file(s) download failed (may be wrong svn path or no internet)<br>\" >> $applogPath/appInstall.log; touch $applogPath/fail.txt; fi";
+			$checkoutCmd = "mkdir $packdir; cd $packdir; if svn --non-interactive --username osxlatitude-edp-read-only --quiet --force co http://osxlatitude-edp.googlecode.com/svn/$svnpath; then echo \"$fname file(s) download finished<br>\"; touch $appsLogPath/Success_$fname.txt; else echo \"$fname file(s) download failed (may be wrong svn path or no internet)<br>\"; touch $appsLogPath/Fail_$fname.txt; fi";
 
-			writeToLog("$applogPath/$fname.sh", "$createStatFile; $checkoutCmd; $endStatFile;");
-			system_call("sh $applogPath/$fname.sh >> $applogPath/appInstall.log &");	
+			writeToLog("$appsLogPath/$fname.sh", "$checkoutCmd;");
+			system_call("sh $appsLogPath/$fname.sh >> $appsLogPath/appInstall.log &");	
 		}
 } 
 ?>
